week2/day5_livekit/requirements.txt
livekit-agents[openai,deepgram,elevenlabs,silero]>=0.8.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
